feat: Create consult endpoint

+++ src/main/resources/db/migration/V1__create_table_payments.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS "payments"."payment"
 (
   internal_payment_id varchar(36) not null,
-  external_payment_id integer,
+  external_payment_id bigint,
   external_id varchar(100) not null,
   payer varchar(100) not null,
   payment_amount decimal(6, 2) not null,
